add flag for GUI
I added a flag for GUI so now you can use --text or --gui. The program
will default to --text if none are entered. Running the bash script
still defaults to --text, so just change it if you need to. I still have
not implemented the actual GUI launching yet, and I'm hoping to get to
that in my next commit to launch a window.