add elis FXscratch and andrews viewer to archive
I added Eli's work that he did on his branch, and I also moved Andrew's old viewer code to archive as well if we need to reference it. This way we can start fresh with the GUI implementation. I am also going to delete Eli's branch. You may need to reclone the repo after this.