create 17x17 and place playable tiles
I got the 17x17 grid to display properly, and also the player tiles in
the 9x9 to work. I have walls being drawn on the board, but I'm running
into issues where they do not overlap the board properly and instead
move all of the grid elements over. I'm also not sure if we should have
a separate class for the Wall or just handle it within Tile.