add init function, modify tile constructor
I added an init function to clean out the start function for increased
readability. I also modified the constructor of the tile to accept an x
an y value (row and column value) and now prints when the tiles are
drawn. We still need a data structure to hold the tiles to work with the
board.