getPaths() method added
Takes in a player and a board object, returns an arrayList of int[] arrays, each of which holds the shortest path from player to each node in that row.  Currently it prints out each row, but that can be taken out.  This will be used by our AI for debugging and making smart moves.